Improvements & bug fixes

Show different colours for each log level (#6)
Merge ServiceAccount and MinioCredentials class #15 
Improve policy sorting to prevent unnecessary updates #22 
Introduce allowed bucket name prefix #31 
Improve service account policy robustness #32 
Catch resources that are defined multiple times #34
Handle service account names longer than 32 characters #35

+++ minio_manager/classes/secrets.py
@@ -6,19 +6,16 @@ from tempfile import NamedTemporaryFile
 
 from minio import Minio, S3Error
 from pykeepass import PyKeePass
+from pykeepass.exceptions import CredentialsError
 
 from minio_manager.classes.config import MinioConfig
+from minio_manager.classes.minio_resources import ServiceAccount
 from minio_manager.utilities import get_env_var, logger
 
 
-class MinioCredentials:
-    def __init__(self, name: str, access_key: str | None = None, secret_key: str | None = None):
-        self.name = name
-        self.access_key = access_key
-        self.secret_key = secret_key
-
-
 class SecretManager:
+    """SecretManager is responsible for managing credentials"""
+
     def __init__(self, config: MinioConfig):
         logger.debug("Initialising SecretManager")
         self._cluster_name = config.name
@@ -31,7 +28,7 @@ class SecretManager:
         self.keepass_group = None
         self.backend_s3 = self.setup_backend_s3()
         self.backend = self.setup_backend()
-        logger.info(f"SecretManager initialised with backend {self.backend_type}")
+        logger.info(f"Secret backend initialised with {self.backend_type}")
 
     def setup_backend_s3(self):
         endpoint = get_env_var("MINIO_MANAGER_S3_ENDPOINT")
@@ -42,7 +39,15 @@ class SecretManager:
         try:
             s3.bucket_exists(self.backend_bucket)
         except S3Error as s3e:
-            logger.critical(f"{s3e.code}: Does the bucket exist? Does the user have the correct permissions?")
+            if s3e.code == "SignatureDoesNotMatch":
+                logger.critical("Invalid secret key provided for the secret backend bucket user.")
+            if s3e.code == "InvalidAccessKeyId":
+                logger.critical("Invalid access key ID provided for the secret backend bucket user.")
+            if s3e.code == "AccessDenied":
+                logger.critical(
+                    "Access denied for the secret backend bucket user. Does the bucket exist, and does the "
+                    "user have the correct permissions to the bucket?"
+                )
             sys.exit(10)
         return s3
 
@@ -53,19 +58,20 @@ class SecretManager:
         method = getattr(self, method_name)
         return method()
 
-    def get_credentials(self, name: str) -> MinioCredentials:
+    def get_credentials(self, name: str, required: bool = False) -> ServiceAccount:
         """Get a password from the configured secret backend.
 
         Args:
-            name: str, the name of the password entry
+            name (str): the name of the password entry
+            required (bool): whether the credentials must exist
 
         Returns: MinioCredentials
         """
         method_name = f"{self.backend_type}_get_credentials"
         method = getattr(self, method_name)
-        return method(name)
+        return method(name, required)
 
-    def set_password(self, credentials: MinioCredentials):
+    def set_password(self, credentials: ServiceAccount):
         method_name = f"{self.backend_type}_set_password"
         method = getattr(self, method_name)
         self.backend_dirty = True
@@ -77,7 +83,7 @@ class SecretManager:
     def dummy_get_credentials(self, name):
         raise NotImplementedError
 
-    def dummy_set_password(self, credentials: MinioCredentials):
+    def dummy_set_password(self, credentials: ServiceAccount):
         raise NotImplementedError
 
     def retrieve_keepass_backend(self) -> PyKeePass:
@@ -111,7 +117,11 @@ class SecretManager:
 
         kp_pass = get_env_var("MINIO_MANAGER_KEEPASS_PASSWORD")
         logger.debug("Opening keepass database")
-        kp = PyKeePass(self.keepass_temp_file_name, password=kp_pass)
+        try:
+            kp = PyKeePass(self.keepass_temp_file_name, password=kp_pass)
+        except CredentialsError:
+            logger.critical("Invalid credentials for Keepass database.")
+            sys.exit(13)
         # noinspection PyTypeChecker
         self.keepass_group = kp.find_groups(path=["s3", self._cluster_name])
         if not self.keepass_group:
@@ -120,34 +130,37 @@ class SecretManager:
         logger.debug("Keepass configured as secret backend")
         return kp
 
-    def keepass_get_credentials(self, name: str) -> MinioCredentials | bool:
+    def keepass_get_credentials(self, name: str, required: bool) -> ServiceAccount:
         """Get a password from the configured Keepass database.
 
         Args:
-            name: str, the name of the password entry
+            name (str): the name of the password entry
+            required (bool): if the entry must exist
 
         Returns:
-            MinioCredentials if found, False if not found
+            ServiceAccount
         """
         logger.debug(f"Finding Keepass entry for {name}")
         entry = self.backend.find_entries(title=name, group=self.keepass_group, first=True)
 
         try:
-            credentials = MinioCredentials(name, entry.username, entry.password)
+            credentials = ServiceAccount(name=name, access_key=entry.username, secret_key=entry.password)
             logger.debug(f"Found access key {credentials.access_key}")
         except AttributeError as ae:
             if not ae.obj:
-                logger.warning(f"Entry for {name} not found!")
-                return MinioCredentials(name=name)
+                if required:
+                    logger.critical(f"Required entry for {name} not found!")
+                    sys.exit(14)
+                return ServiceAccount(name=name)
             logger.critical(f"Unhandled exception: {ae}")
         else:
             return credentials
 
-    def keepass_set_password(self, credentials: MinioCredentials):
+    def keepass_set_password(self, credentials: ServiceAccount):
         """Set the password for the given credentials.
 
         Args:
-            credentials: MinioCredentials
+            credentials (ServiceAccount): the credentials to set
         """
         logger.info(f"Creating Keepass entry for {credentials.access_key}")
         self.backend.add_entry(
@@ -159,7 +172,8 @@ class SecretManager:
 
     def cleanup(self):
         if not self.backend_dirty:
-            Path(self.keepass_temp_file_name).unlink()
+            if self.keepass_temp_file_name:
+                Path(self.keepass_temp_file_name).unlink()
             return
 
         # If we have dirty back-ends, we want to ensure they are saved before exiting.
